src/requirements.txt
aiohttp
kafka-python
psycopg2-binary
pyyaml
